src/ui/requirements.txt
CairoSVG==2.7.1
chess==1.11.1
numpy==1.21.5
Pillow==11.0.0
pygame==2.6.1
